Move the high-level API to bazil.org/fuse/fs.
The top-level now contains things related to the raw kernel protocol:
func Mount and type Conn that are about the communication itself,
XxxRequest and XxxResponse types, and types used for communication
such as Attr.

fs now contains the Serve function, and anything that keeps in-memory
state as part of the serving. It contains interfaces that the Serve
API uses, including interfaces FS, Node, and Handle.

This will hopefully make the documentation more browseable.