Add NodeRef, allowing Nodes be recognized across calls
Without this, a simple looping write+read workload of ~10MB sees about
1.3x times the Lookups and 5x Forgets, as the nodes are not recognized
between multiple Create/Lookup calls.

Unfortunately, this is not easy to test in isolation; triggering the
bad behaviour seems to require an EntryValid timeout (or something
like that; something causes the kernel to decide to validate the
cache) and a second Lookup that creates a new NodeID. This has so far
not been easily reproducible in a timely fashion.