Support Getxattr
Needed by Camlistore. This implementation is from-scratch and does not
use the work in Camlistore's third_party directory, opting instead to
stay closer in style of the existing code.

Noteworthy difference: Camlistore picked ENOATTR as the error for all
platform Getxattr calls to return. As this seems to be only used by OS
X, I'm opting to use ENODATA instead, consider OS X the odd one out,
and translate ENODATA to ENOATTR on darwin.