Refactor TestStatfs
On Linux, waitForMount is not needed; there mounting is synchronous
and if Mount returns success, all accesses will wait until Serve gets
to them.

On OS X, mount wants to perform the InitRequest handshake
synchronously, so current code just starts it off in a child process
and forgets about it.

This will be cleaned up and made more robust later.