Refactor TestMkdir
Store a copy of the fuse.MkdirRequest so there's no worries about
lifetime or request object reuse. Avoid pointers to minimize
boilerplate in caller; zero value should be distinct enough to make
comparisons fail if request was never seen.

Zero the fuse.Request Header as uninteresting and hard to compare.