Mount can now report when new mount is visible
On Linux, this has always been synchronous, and the new Ready channel
is closed immediately.

OSXFUSE forces us to jump through hoops to know when the mount is
visible. Mount now keeps track of the mount_osxfusefs child process,
and once the child exits, signals the app (via the Ready channel) that
the mount is either visible or has failed. This means there are now
two sources of errors: fuse.Conn.MountError and fs.Serve. This is
ugly, but reflects what OSXFUSE forces on us.

This removes the need to poll for the mountpoint (and the possibility
for false positives/false negatives), and also removes those extra
operations from every debug log.