Pretty logging for error messages
Instead of just stringifying the error, format it so that both the
errno and the freeform error message (if any) are visible. For
example: "EIO: connection timed out".

This makes sense because a FUSE file system developer needs to be
acutely aware of the errnos used, and now the work can be done in just
one place, and the developer can memorize the identifiers used in the
source, instead of the long form descriptions of the errors.

For structured logging, as encoding/json and others would marshal many
custom error types as just "{}" (struct with no exported fields),
preprocess the error into two fields:

- Errno: the errno constant as a string, for example "ENOENT"; always
  set for errors

- Error: the message returned from Error(); may not be present