hellofs reports mount errors on OS X
On OS X, the mount helper can exit with an error, but is not
guaranteed to exit until after the Serve loop has started running.
We need to catch the error after Serve bails out.

This was triggered by "./hellofs /does-not-exist" failing to mount
anything, but exiting succesfully and not giving an error message.

Also close the fuse.Conn cleanly.

Thanks to James Gray <james@james4k.com> for discovering the bug.