OS X: Limit I/O size to 4kB, fixes large writes
Add back `-o iosize=4096` to `mount_osxfusefs` options.

Without this, osxfuse-kext will try to send large writes to userspace,
and as the `read(2)` buffer is not large enough to copy them to
userspace, this leads to a FUSE layer failure, and the kernel
connection closing.

This is a temporary workaround. Larger writes are desirable, but first
we need better buffer management.

Thanks to Michael Frister <michael+git@frister.net> and Eric Drechsel
<eric@pdxhub.org>.

Fixes: #12
Fixes: #18