fs/fstestutil: write self-contained Linux getMountInfo
Avoids dependence on github.com/artyom/mtab,
which has no LICENSE file, and makes the file shorter anyway.
This is a trivial problem; it doesn't merit the complexity
of depending on other code.

The lack of license in the current dependency is a problem
for users trying to do everything by the book.