Getxattr and friends now return ErrNoXattr
There is no consistency in ENODATA/ENOATTR across platforms, and
supporting FreeBSD and OpenBSD means we can't even assume
syscall.ENODATA exists.

Move from an approach of returning ENODATA everywhere and translating
it to the correct value to an approach where each platform gets to
define the value of the ErrNoXattr constant. This is better than
pretending that all platforms have ENODATA.

Filesystems intending to be platform independent should adjust
Getxattr etc methods to return ErrNoXattr instead of ENODATA.