Change Node.Attr to take a pointer to fuse.Attr to fill, can now set Nlink=0
This allows Serve to both fill in default values, and let the Node
change them as it wishes.

Before this, Nlink==0 was confused for the default value, and Serve
overrode it. Link count 0 is useful to be able to communicate unlinked
files.

This is a small but pervasive API change. Apologies for the trouble.

fixes https://github.com/bazillion/fuse/issues/66