Don't consider fusermount errors fatal if exit status is still 0
The primary effect of this is to let users mount FUSE filesystems even
when they're not members of group fuse. Tested on Ubuntu 14.04; other
distros probably behave similarly.

In an attempt to make the average experience more pleasant, hide the
particular fusermount error message resulting from not being a member
of group fuse; it seems to in no way hinder the actual mount.