API change: Attr method takes Context, returns error
This allows e.g. network operations at that time. This was possible in
the Getattr method already, but that's only called specifically for
FUSE Getattr requests; there are many other times when attributes are
needed, such as Lookup, Create, Mkdir, Link and such request handling.

Most of the change can be automated with a regexp monster like

    find -name '*.go' | xargs perl -0777 -p -i -e 's/^(func \(.*?\) Attr\()([^,]*\*fuse\.Attr\)) (\{\n(?:(?:\s.*)?\n)*)(}\n)/$1ctx context.Context, $2 error $3\treturn nil\n$4/mg' --
    goimports -w .