Unify /dev/fuse write error handling
bugShortKernelWrite now triggers for both respond and respondData. It
does not trigger when the write claims failure -- that's not a short
write. By now, the bugs it detected should be fixed, with the proper
MaxWrite handling.

A separate error, bugKernelWriteError, triggers when writing a
response gives an error. This is somewhat an unknown territory, and
seems to trigger every now and then, for unknown reasons. See
https://github.com/bazil/fuse/issues/90 .