Fixed a double-close bug in mount on Linux.
The socket file descriptors were previously closed by both a direct call
to syscall.Close and by a deferred call to os.File.Close. Now we close
them only once.

In the common case this would result in error EBADF from the second pair
of calls, which was ignored. However, there was a race that would come
up when the process was quite busy, especially with GOMAXPROCS > 1:

 *  Goroutine 1: mount does File.Close(4).
 *  Goroutine 2: os.Open returns a file with Fd() == 4.
 *  Goroutine 1: mount does syscall.Close(4)
 *  Goroutine 2: Interacting with the new file always returns EBADF.

I discovered this when attempting to parallelize a set of tests that
make heavy use of fuse. The double-close caused chaos, with EBADF errors
all over the place in unlikely situations. After staring at strace
output for a long time, I figured out this is what was happening.