Refactor outgoing /dev/fuse message construction
We want to get rid of every kernel message type embedding outHeader,
as we need more flexibility in manipulating them. The response to
Create is really outHeader+entryOut+openOut, and with FUSE protocol
version 7.9 the part in the middle grows. This is too difficult to
manage with the struct design that was before.

For easier software archeology, this change will be done in a series
of commits, each one changing one type.