Support custom Context factories
fs.Config field Context can now be used to create contexts that e.g.
remember app-specific details. Putting it in Config lets us avoid an
API change.

We use a constructor function and not a single context.Context value
to e.g. allow custom timeouts, which obviously can't measure time from
all the way from when the application started.

This may not be the final form of how we use top-level contexts, but
at least this change is not intrusive to those who don't need the
functionality right now.