Stop over-allocating receive buffers for large writes.
The maxWrite constant used to size buffers was 16 MiB, even though
InitRequest.Respond always tells the kernel we want writes of 128 KiB.
Caveat: mount_darwin.go implies that on darwin we were still telling
the kernel 16 MiB, but experimentally it was capping this to 1 MiB.

Now use exactly those sizes on those operating systems. On FreeBSD
continue to use 128 KiB, because I'm not sure what it does.

Allocating and zeroing buffers in allocMessage is a large source of CPU
usage in a read-heavy load; I believe this change will help there.