Replaced sync.Pool with a dead-simple freelist.
The behavior of sync.Pool is hard to predict, because it interacts with
garbage collection. It goes to considerable effort to not retain items
in its freelist when there is memory pressure, but this can cause
pathological behavior where we allocate over and over again when in
theory we need only one message in flight
(cf. http://golang.org/issue/11834).

This change causes allocMessage to go from allocating hundreds of MB per
second in my read throughput benchmark to not being in the profile at
all.

The downside is that the freelist grows to a high water mark of messages
that were needed concurrently, and then stays there. I suspect this is
not a problem in practice. If it is, then we could either reset it
periodically, or use additive increase/multiplicative decrease
(cf. https://goo.gl/uk7JS9) on its capacity.