Fill SetattrResponse.Attr by default
Without this, if the Setattr method didn't explicitly fill that, we
told the kernel that it's inode 0, and so on. It's currently unclear
whether that had any real implications, but this is better (at least
it should avoid an extra Getattr roundtrip), and symmetrical with
Lookup/Create/Mkdir/Symlink doing the same for LookupResponse.