Fix OSXFUSE 2.x detection logic
Previous logic only worked when the 2.x kext was NOT loaded; that is,
right after boot. We only got it right if we needed to load the kext,
and the load helper binary wasn't found.

Instead, just probe for the mount helper binary, before doing anything
else. This costs an extra `stat(2)`, but keeps the code simple.