Split TestInvalidateNodeData into separate attribute and data tests
This makes it a little easier to grok the differences between Linux
and OSXFUSE.

OSXFUSE 3.0.6 sends an immediate Getattr on invalidate, Linux just
marks the cached entry as expired. On Linux, reading beyond the
(cached, perhaps stale) EOF causes a Getattr call to update the size.
Otherwise, Linux will wait for the first `stat(2)` or such need to
perform the Getattr.