Remove incorrect TODO note
MOUNT_FUSEFS_DAEMON_PATH is used only for three things:

1. derive default value for FSName
2. derive default value for VolumeName
3. short list of hardcoded names

The first two already have perfectly good means of setting the values.

The last one would need changes in OSXFUSE upstream per filesystem, so
it's not very likely to be used anyway, and the basename of the
executable should work just fine for the detection anyway.

Resources:

- https://code.google.com/p/macfuse/wiki/OPTIONS
- https://github.com/osxfuse/support/blob/664d8babd2aca73e074ceb5053336343e28c5329/osxfusefs/osxfusefs-Info.plist#L24-L117