fuse: added darwin specific mount options.
In particular:

NoAppleXattr - to avoid apple specific xattrs
NoAppleDouble - to avoid using ._<filename> files to store xattrs.
DaemonTimeout - to change the 'mount point is dead' timeout.

Also changed the SubType option for OSX to use the correct flag.

Also added an environment variable, MOUNT_OSXFUSE_CALL_BY_LIB, when executing the
mount helper.  OSXFUSE seems to have changed that from MOUNT_FUSEFS_CALL_BY_LIB when
moving to major version 3 so we have to define both when calling the helper to work
with both versions 2 and 3.