Use reusable read buffer to improve performance
On my simple, sample filesystem, this speeds up my benchmark
(a series of git clones) by a factor of 5x-10x.

The vast majority of the time savings is from
reduced allocation and GC of giant buffers.

Some representative runtime.MemStats fields after a run, before:

TotalAlloc=302219702256 Mallocs=343942 Frees=342827 GCCPUFraction=0.199405

After:

TotalAlloc=42925688 Mallocs=194497 Frees=193819 GCCPUFraction=0.000480

Note that ReadRequest is normally called from
exactly one goroutine, as part of a loop in Server.serve,
so switching rio to a sync.Mutex will actually help performance
(sync.RWMutex has non-trivial additional overhead).
It also makes the code simpler and easier to reason about.

This does introduce expensive copies for large buffers.
I experimented with having three tiers of buffer sizes instead.
In that scenario, small buffers are pooled (as in this commit),
medium buffers are allocated and the data copied over (as in this commit),
but very large buffers take over c.readbuf instead of copying it,
and allocate a new c.readbuf in its place.
In my experiments, in order for it to be worth allocating a full sized buffer,
the cutoff to count as "very large" has to be very large indeed,
and I don't see any evidence that such messages are common enough
to warrant the extra code (if indeed such messages exist in practice at all).