Disable test parallelism, for now
Empirically, running tests one by one decreases the likelihood of
hanging in epoll significantly (see
https://github.com/golang/go/issues/21014 ). Perhaps this can be
re-enabled after that mess is sorted out (and generally triggering
bugs in tests is good!). Or maybe not; the way go test logs parallel
tests isn't very readable for the style of tests we have: long
interconnected sequences of messages that are crucial for debugging,
and a high percentage of bugs being hangs.