Add handle-specific attr and fsync interfaces
Getattr and Setattr requests can either happen on a fs.Node or an
fs.Handle (depending on the call that initiated it e.g. truncate vs
ftruncate or stat vs fstat). Add HandleGetattrer and HandleSetattrer so
implementations can provide handle-specific overrides when those
operations are performed on an open file descriptor.

Also create HandleFsyncer and deprecate NodeFsyncer to resolve the TODO.

Resolves #225.