Use spawntest for TestInterrupt, avoid SIGIO
It seems SIGIO worked earlier to trigger EINTR only because the child
process was so simple. As soon as it contained multiple active
goroutines, the test got flaky (theory: behavior depended on which
thread received the signal). Port it to use SIGSTOP/SIGCONT pair,
which seems to trigger EINTR much more reliably.

Throw away the macOS-specific kludges that just kill the process,
don't trigger EINTR. If that gets resurrected, it'll more
appropriately live as a separate test, the shared code was more
accidental than useful. And if macOS can't run this test, we'll just
add a conditional t.Skip at that time.