Node and Handle methods can now return syscall.Errno directly
This removes the need for fuse.Errno wrappers like fuse.ENOENT.
Existing values will be kept for a while for backwards
compatilibility, but now new ones will be added.

Custom errors can now implement fuse.ErrorNumber via wrapping, for
more information see go1.13 changes and errors.As.

See fuse.ToErrno for details.

Also add ENAMETOOLONG to errnos we know the short name for, because I
use it in tests a lot and like the pretty debug output -- but do not
expose a fuse.ENAMETOOLONG.

Thanks to Pallav Agarwal <pallavagarwal07@gmail.com> for the perfect
nudge in the right direction.