Support time_gran and max_pages fuse options
This updated the supported FUSE protocol from 7.12 to 7.28.

The time_gran field can be set to a multiple of one nanosecond, if the
timestamps of files has less than nanosecond precision. This avoids
mismatches when the kernel compares timestamps.

The max_pages feature raises the maximum size of a FUSE message from 32
to 256 pages on Linux >= 4.20. This affects both reads and writes.
To be able to read more than 128 KB in a single ReadOp, Direct I/O must
also be used, since the kernel otherwise splits the read into 128 KB
chunks and requests them asynchronously (if max_readahead is at least
128 KB, otherwise max_readahead is used, or 4 KB if it is disabled).