Fix TestFeatures odd delay
It started occurring on Linux too. It seems fusermount (sometimes?)
access the mountpoint, e.g. stats it, and the kernel waited for the
FUSE connection to respond for a while. Close the connection first so
it's a dead mount, and fusermount proceeds immediately.

This also fixes the same delay observed on FreeBSD.